Added separate classes for Calculate water and light in server.
