+++ src/main/java/se/myhappyplants/server/controller/Main.java
@@ -1,8 +1,10 @@
 package se.myhappyplants.server.controller;
 
-import se.myhappyplants.server.model.repository.DBPlantRepository;
-import se.myhappyplants.server.model.repository.UserPlantRepository;
-import se.myhappyplants.server.model.repository.UserRepository;
+import se.myhappyplants.server.model.LightCalculator;
+import se.myhappyplants.server.model.WaterCalculator;
+import se.myhappyplants.server.repository.PlantRepository;
+import se.myhappyplants.server.repository.UserPlantRepository;
+import se.myhappyplants.server.repository.UserRepository;
 
 /**
  * Created by: Frida Jacobson, Eric Simonson, Anton Holm, Linn Borgström, Christopher O'Driscoll
@@ -10,7 +12,9 @@ import se.myhappyplants.server.model.repository.UserRepository;
 public class Main {
     public static void main(String[] args) throws Exception {
         Controller controller = new Controller();
-        new Server(2555, new UserRepository(), new UserPlantRepository(controller), new DBPlantRepository(controller), controller);
+        LightCalculator lightCalculator = new LightCalculator();
+        WaterCalculator waterCalculator = new WaterCalculator();
+        new Server(2555, new UserRepository(), new UserPlantRepository(lightCalculator, waterCalculator), new PlantRepository(lightCalculator, waterCalculator), controller);
 
     }
 }
